Rename module files
