+++ lib/hook.js
