+++ lib/plugins/page-visibility-tracker.js
@@ -16,10 +16,10 @@
 
 
 import {NULL_DIMENSION} from '../constants';
-import Hook from '../hooks';
+import Hook from '../hook';
 import provide from '../provide';
 import Session from '../session';
-import Store from '../storage';
+import Store from '../store';
 import {plugins, trackUsage} from '../usage';
 import {assign, createFieldsObj, isObject, now, uuid} from '../utilities';
