+++ lib/plugins/url-change-tracker.js
@@ -15,7 +15,7 @@
  */
 
 
-import Hook from '../hooks';
+import Hook from '../hook';
 import provide from '../provide';
 import {plugins, trackUsage} from '../usage';
 import {assign, createFieldsObj} from '../utilities';
