+++ lib/session.js
@@ -15,8 +15,8 @@
  */
 
 
-import Hook from './hooks';
-import Store from './storage';
+import Hook from './hook';
+import Store from './store';
 import {now} from './utilities';
